# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Funding Application Platform API
  version: 1.0.0
  description: |
    RESTful API for the Funding Application Submission & Assessment Platform.

    This API enables:
    - **Applicants** to submit applications with uploaded files and confirmations
    - **Coordinators** to manage funding calls, assign assessors, and view master results
    - **Assessors** to score assigned applications against defined criteria

    ## Authentication

    All endpoints except public call listings require JWT Bearer authentication.
    Service-to-service communication uses API key authentication.

    ## Rate Limiting

    - Standard endpoints: 100 requests/minute per user
    - File upload endpoints: 10 requests/minute per user
    - Export endpoints: 5 requests/minute per user

    ## Timezone Handling

    All timestamps are stored in UTC and returned in ISO 8601 format.
    Deadlines are enforced in Europe/London timezone.
  contact:
    name: Platform Support
    email: support@fundingplatform.example.com
    url: https://fundingplatform.example.com/support
  license:
    name: Proprietary
    url: https://fundingplatform.example.com/license
  x-logo:
    url: https://fundingplatform.example.com/logo.png
    altText: Funding Platform Logo
servers:
  - url: https://api.fundingplatform.example.com/api/v1
    description: Production server
  - url: https://staging-api.fundingplatform.example.com/api/v1
    description: Staging server
...
security:
  - BearerAuth: []
paths:
...
  /results:
    get:
      tags:
        - Results
      summary: Get master results
      description: |
        Returns aggregated assessment results for a funding call.

        Includes per-assessor scores, calculated totals/averages, and variance flags.

        **Required role:** Coordinator (results are not visible to assessors)
      operationId: getMasterResults
      parameters:
        - name: call_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: sort
          in: query
          schema:
            type: string
            enum: [total_score, average_score, variance, reference]
            default: total_score
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Master results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterResultsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
...
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              details: null
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "FORBIDDEN"
              message: "You do not have permission to perform this action"
              details:
                required_role: "coordinator"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
              details: null
...
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Request validation failed"
            details:
              type: object
              nullable: true
              description: Additional error details
              additionalProperties: true

    # --------------------------------------------------------------------------
    # PAGINATION SCHEMA
    # --------------------------------------------------------------------------
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_items:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8

    # --------------------------------------------------------------------------
    # AUTHENTICATION SCHEMAS
    # --------------------------------------------------------------------------
...
    Score:
      type: object
      properties:
        criterion_id:
          type: string
        criterion_name:
          type: string
        score:
          type: integer
        max_points:
          type: integer
        weight:
          type: number
        comment:
          type: string
...
    MasterResultsResponse:
      type: object
      properties:
        call_id:
          type: string
          format: uuid
        call_name:
          type: string
        summary:
          type: object
          properties:
            total_applications:
              type: integer
            fully_assessed:
              type: integer
            high_variance_count:
              type: integer
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApplicationResult'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ApplicationResult:
      type: object
      properties:
        application_id:
          type: string
          format: uuid
        reference:
          type: string
        applicant_name:
          type: string
        applicant_organisation:
          type: string
        assessments:
          type: array
          items:
            type: object
            properties:
              assessor_id:
                type: string
                format: uuid
              assessor_name:
                type: string
              total_score:
                type: number
              scores:
                type: array
                items:
                  $ref: '#/components/schemas/Score'
              overall_comment:
                type: string
              submitted_at:
                type: string
                format: date-time
        aggregates:
          type: object
          properties:
            total_score_sum:
              type: number
            total_score_average:
              type: number
            weighted_average:
              type: number
            by_criterion:
              type: array
              items:
                type: object
                properties:
                  criterion_id:
                    type: string
                  criterion_name:
                    type: string
                  average_score:
                    type: number
                  max_points:
                    type: integer
                  variance:
                    type: number
        variance_flags:
          type: array
          items:
            type: object
            properties:
              criterion_id:
                type: string
              criterion_name:
                type: string
              variance_percentage:
                type: number
              min_score:
                type: number
              max_score:
                type: number
        rank:
          type: integer
          description: Rank by average score